Split CLI commands into their own files.
