+++ fiona/fio/distrib.py
@@ -1,0 +1,38 @@
+"""$ fio distrib"""
+
+
+import json
+import logging
+
+import click
+import cligj
+
+from fiona.fio import helpers
+
+
+@click.command(short_help="Distribute features from a collection")
+@cligj.use_rs_opt
+@click.pass_context
+def distrib(ctx, use_rs):
+
+    """Print the features of GeoJSON objects read from stdin."""
+
+    verbosity = (ctx.obj and ctx.obj['verbosity']) or 2
+    logger = logging.getLogger('fio')
+    stdin = click.get_text_stream('stdin')
+    try:
+        source = helpers.obj_gen(stdin)
+        for i, obj in enumerate(source):
+            obj_id = obj.get('id', 'collection:' + str(i))
+            features = obj.get('features') or [obj]
+            for j, feat in enumerate(features):
+                if obj.get('type') == 'FeatureCollection':
+                    feat['parent'] = obj_id
+                feat_id = feat.get('id', 'feature:' + str(i))
+                feat['id'] = feat_id
+                if use_rs:
+                    click.echo(u'\u001e', nl=False)
+                click.echo(json.dumps(feat))
+    except Exception:
+        logger.exception("Exception caught during processing")
+        raise click.Abort()
