+++ fiona/fio/filter.py
@@ -1,3 +1,6 @@
+"""$ fio filter"""
+
+
 import json
 import logging
 import math
