+++ setup.py
@@ -208,15 +208,15 @@ setup_args = dict(
         [fiona.fio_commands]
         bounds=fiona.fio.bounds:bounds
         cat=fiona.fio.cat:cat
-        collect=fiona.fio.cat:collect
-        distrib=fiona.fio.cat:distrib
-        dump=fiona.fio.cat:dump
-        env=fiona.fio.info:env
-        info=fiona.fio.info:info
-        insp=fiona.fio.info:insp
-        load=fiona.fio.cat:load
-        ls=fiona.fio.ls:ls
+        collect=fiona.fio.collect:collect
+        distrib=fiona.fio.distrib:distrib
+        dump=fiona.fio.dump:dump
+        env=fiona.fio.env:env
         filter=fiona.fio.filter:filter
+        info=fiona.fio.info:info
+        insp=fiona.fio.insp:insp
+        load=fiona.fio.load:load
+        ls=fiona.fio.ls:ls
         ''',
     install_requires=requirements,
     tests_require=['nose'],
